+++ tests/testthat/test-compareModified.R
@@ -1,6 +1,6 @@
 with_demoRepo({
   if (Sys.getenv("METWORX_VERSION") != "") {
-    test_that("compareModifiedFigures doesn't return anything", {
+    test_that("compareModified doesn't return anything", {
       
       expect_message({
         compareModified(fs::path_abs("deliv/figure/example-pdf1.pdf"))